fix: standardize JOD money precision to three decimals
Widens cost_cents/sell_cents/unit_cents from integer to decimal(15,1)
on event_transport, event_invoice_items, and invoice_lines — one more
fractional digit on the stored cents value, which is one more decimal
place on the currency amount once divided by 100. Existing whole-cent
values are already valid decimal(15,1) values, so nothing already
saved changes; down() reverts the column types.

Money::forScreen/forDocument/abbreviated now accept int|float, and
forDocument takes an explicit $decimals (default 2) for the screens
that opt into a third decimal. Invoice totals: subtotal/net/total/
outstanding follow the widened line precision; fee and tax continue
rounding to a whole cent, matching how they always behaved.

+++ app/Models/Invoice.php
@@ -134,7 +134,7 @@ class Invoice extends Model
 
     /* ── money ── */
 
-    public function subtotalCents(): int
+    public function subtotalCents(): float
     {
         return $this->lines->sum(fn (InvoiceLine $l) => $l->amountCents());
     }
@@ -152,7 +152,7 @@ class Invoice extends Model
     }
 
     /** What is being charged before tax: the work plus the fee on it. */
-    public function netCents(): int
+    public function netCents(): float
     {
         return $this->subtotalCents() + $this->feeCents();
     }
@@ -168,12 +168,12 @@ class Invoice extends Model
         return (int) round($this->netCents() * $this->tax_pct / 100);
     }
 
-    public function totalCents(): int
+    public function totalCents(): float
     {
         return $this->netCents() + $this->taxCents();
     }
 
-    public function outstandingCents(): int
+    public function outstandingCents(): float
     {
         return max(0, $this->totalCents() - $this->paid_cents);
     }
